chatbot4/newcChatbot/requirements.txt

Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
pymssql==2.2.2
google-cloud-dialogflow==2.14.1
